+++ src/Oro/Bundle/ShoppingListBundle/Manager/ShoppingListOwnerManager.php
@@ -3,6 +3,7 @@
 namespace Oro\Bundle\ShoppingListBundle\Manager;
 
 use Doctrine\Common\Persistence\ManagerRegistry;
+use Doctrine\ORM\AbstractQuery;
 use Doctrine\ORM\EntityRepository;
 use Oro\Bundle\CustomerBundle\Entity\CustomerUser;
 use Oro\Bundle\EntityConfigBundle\Provider\ConfigProvider;
@@ -50,6 +51,7 @@ class ShoppingListOwnerManager
      */
     public function setOwner($ownerId, ShoppingList $shoppingList)
     {
+        /** @var CustomerUser $user */
         $user = $this->registry->getRepository(CustomerUser::class)->find($ownerId);
         if (null === $user) {
             throw new \InvalidArgumentException(sprintf("User with id=%s not exists", $ownerId));
@@ -58,7 +60,9 @@ class ShoppingListOwnerManager
             return;
         }
         if ($this->isUserAssignable($ownerId)) {
+            $this->assignLineItems($shoppingList, $user);
             $shoppingList->setCustomerUser($user);
+
             $this->registry->getManagerForClass(ShoppingList::class)->flush();
         } else {
             throw new AccessDeniedException();
@@ -75,8 +79,9 @@ class ShoppingListOwnerManager
         $repository = $this->registry->getRepository(CustomerUser::class);
         $qb = $repository
             ->createQueryBuilder('user')
-            ->where("user.id = :id")
-            ->setParameter("id", $id);
+            ->select('user.id')
+            ->where('user.id = :id')
+            ->setParameter('id', $id);
 
         $query = $this->aclHelper->apply(
             $qb,
@@ -88,8 +93,17 @@ class ShoppingListOwnerManager
             ]
         );
 
-        $owner = $query->getOneOrNullResult();
+        return null !== $query->getOneOrNullResult(AbstractQuery::HYDRATE_SINGLE_SCALAR);
+    }
 
-        return null !== $owner;
+    /**
+     * @param ShoppingList $shoppingList
+     * @param CustomerUser $user
+     */
+    private function assignLineItems(ShoppingList $shoppingList, CustomerUser $user): void
+    {
+        foreach ($shoppingList->getLineItems() as $lineItem) {
+            $lineItem->setCustomerUser($user);
+        }
     }
 }