+++ src/Oro/Bundle/VisibilityBundle/Visibility/Cache/Product/CustomerProductResolvedCacheBuilder.php
@@ -13,6 +13,9 @@ use Oro\Bundle\VisibilityBundle\Entity\VisibilityResolved\CustomerProductVisibil
 use Oro\Bundle\VisibilityBundle\Entity\VisibilityResolved\Repository\CustomerProductRepository;
 use Oro\Bundle\VisibilityBundle\Visibility\Cache\ProductCaseCacheBuilderInterface;
 
+/**
+ * Product visibility cache builder based on customer.
+ */
 class CustomerProductResolvedCacheBuilder extends AbstractResolvedCacheBuilder implements
     ProductCaseCacheBuilderInterface
 {
@@ -72,7 +75,7 @@ class CustomerProductResolvedCacheBuilder extends AbstractResolvedCacheBuilder i
         }
 
         $this->executeDbQuery($er, $insert, $delete, $update, $where);
-        $this->triggerProductReindexation($product, $scope->getWebsite());
+        $this->triggerProductReindexation($product, $scope->getWebsite(), false);
     }
 
     /**
@@ -86,7 +89,7 @@ class CustomerProductResolvedCacheBuilder extends AbstractResolvedCacheBuilder i
     /**
      * {@inheritdoc}
      */
-    public function productCategoryChanged(Product $product)
+    public function productCategoryChanged(Product $product, bool $scheduleReindex)
     {
         $category = $this->registry
             ->getManagerForClass('OroCatalogBundle:Category')
@@ -103,7 +106,7 @@ class CustomerProductResolvedCacheBuilder extends AbstractResolvedCacheBuilder i
         $this->getRepository()->deleteByProduct($product);
         $this->getRepository()->insertByProduct($this->insertExecutor, $product, $category);
 
-        $this->triggerProductReindexation($product);
+        $this->triggerProductReindexation($product, null, $scheduleReindex);
     }
 
     /**